LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-es-ES.html

<!DOCTYPE html><!-- webkit-test-runner [ language=es-ES ] -->
<html>
<head>
<script src="../../../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "es-es");
</script>
</body>
</html>
